+++ common/src/python/keys/keys.py
@@ -49,6 +49,8 @@ class DefaultValues:
     UDS_MODULE = "UDS"
     MDS_MODULE = "MDS"
     BDS_MODULE = "BDS"
+    NP_MODULE = "NP"
+    MLST_MODULE = "MLST"
     GEARBOT_USER_ID = "nacc-flywheel-gear@uw.edu"
     NACC_GROUP_ID = "nacc"
     METADATA_PRJ_LBL = "metadata"
@@ -123,6 +125,9 @@ class SysErrorCodes:
     DUPLICATE_VISIT = "preprocess-022"
     LOWER_VISITNUM = "preprocess-023"
     MISSING_SUBMISSION_STATUS = "preprocess-024"
+    CLINICAL_FORM_REQUIRED = "preprocess-025"
+    DEATH_DATE_MISMATCH = "preprocess-026"
+    AUTOPSY_NP_INVALID = "preprocess-027"
 
 
 class PreprocessingChecks:
@@ -134,6 +139,8 @@ class PreprocessingChecks:
     IVP = "ivp"
     UDSV4_IVP = "udsv4-ivp"
     VISIT_CONFLICT = "visit-conflict"
+    CLINICAL_FORMS = "clinical-forms"
+    NP_MLST_RESTRICTIONS = "np-mlst-restrictions"
 
     @classmethod
     def is_check_defined(cls, check: str) -> bool: